+++ scripts/ci/test_static.sh
@@ -4,9 +4,15 @@ set -e
 
 . $(cd $(dirname $0); pwd)/artifacts.sh
 
+# remove this by public release
 pip install ./privates/azure_mgmt_containerservice-2.0.0-py2.py3-none-any.whl
+
+ls -la $share_folder/build
+
+ALL_MODULES=`find $share_folder/build/ -name "*.whl"`
+
 pip install pylint flake8
-pip install azure-cli-fulltest -f $share_folder/build
+pip install $ALL_MODULES
 
 echo '=== List installed packages'
 pip freeze
@@ -33,6 +39,9 @@ python -m automation.commandlint.run
 
 echo "Verify readme history"
 python -m automation.tests.verify_readme_history
+
+echo "Verify package versions"
+python -m automation.tests.verify_package_versions
 
 echo "Verify default modules"
 azdev verify default-modules $share_folder/build