+++ src/azure-cli-core/azure/cli/core/sdk/util.py
@@ -89,7 +89,7 @@ class CommandGroup(object):
                     no_wait_param=no_wait_param)
 
     def custom_command(self, name, custom_func_name, confirmation=None,
-                       exception_handler=None, deprecate_info=None, no_wait_param=None):
+                       exception_handler=None, deprecate_info=None, no_wait_param=None, table_transformer=None):
         cli_command(self._scope,
                     '{} {}'.format(self._group_name, name),
                     self._custom_path.format(custom_func_name),
@@ -97,10 +97,11 @@ class CommandGroup(object):
                     confirmation=confirmation,
                     deprecate_info=deprecate_info,
                     exception_handler=exception_handler or self._exception_handler,
-                    no_wait_param=no_wait_param)
+                    no_wait_param=no_wait_param,
+                    table_transformer=table_transformer)
 
     def generic_update_command(self, name, getter_op, setter_op, custom_func_name=None,
-                               setter_arg_name='parameters', no_wait_param=None):
+                               setter_arg_name='parameters', no_wait_param=None, **kwargs):
         if custom_func_name:
             custom_function_op = self._custom_path.format(custom_func_name)
         else:
@@ -114,7 +115,7 @@ class CommandGroup(object):
             factory=self._client_factory,
             custom_function_op=custom_function_op,
             setter_arg_name=setter_arg_name,
-            no_wait_param=no_wait_param)
+            no_wait_param=no_wait_param, **kwargs)
 
 
 # PARAMETERS UTILITIES
@@ -155,6 +156,10 @@ class ParametersContext(object):
 
     def register(self, argument_name, options_list, **kwargs):
         register_cli_argument(self._commmand, argument_name, options_list=options_list, **kwargs)
+
+    def extra(self, argument_name, **kwargs):
+        from azure.cli.core.commands import register_extra_cli_argument
+        register_extra_cli_argument(self._commmand, argument_name, **kwargs)
 
     def expand(self, argument_name, model_type, group_name=None, patches=None):
         # TODO: