+++ src/command_modules/azure-cli-monitor/azure/cli/command_modules/monitor/validators.py
@@ -109,3 +109,23 @@ def _validate_tag(string):
         comps = string.split('=', 1)
         result = {comps[0]: comps[1]} if len(comps) > 1 else {string: ''}
     return result
+
+
+def process_action_group_detail_for_creation(namespace):
+    from azure.mgmt.monitor.models import ActionGroupResource, EmailReceiver, SmsReceiver, WebhookReceiver
+
+    _validate_tags(namespace)
+
+    ns = vars(namespace)
+    name = ns['action_group_name']
+    receivers = ns.pop('receivers') or []
+    action_group_resource_properties = {
+        'location': 'global',  # as of now, 'global' is the only available location for action group
+        'group_short_name': ns.pop('short_name') or name[:12],  # '12' is the short name length limitation
+        'email_receivers': [r for r in receivers if isinstance(r, EmailReceiver)],
+        'sms_receivers': [r for r in receivers if isinstance(r, SmsReceiver)],
+        'webhook_receivers': [r for r in receivers if isinstance(r, WebhookReceiver)],
+        'tags': ns.get('tags') or None
+    }
+
+    ns['action_group'] = ActionGroupResource(**action_group_resource_properties)