+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+
+0.1.19
+++++++
+* webapp: fix a bug that downloaded log file might be invalid
+
 0.1.18 (2017-10-09)
 +++++++++++++++++++
 * webapp: added generic update with new command: 'az webapp update'