+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.0.18
+++++++
+* Minor fixes
+
 2.0.17 (2017-10-09)
 +++++++++++++++++++
 * File share snapshot