+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+
+2.0.17
+++++++
+* `vmss create`: expose '--accelerated-networking'
+
 2.0.16 (2017-10-09)
 +++++++++++++++++++
 * `vm show`: fix a bug when using '-d' crashes on missing private ip addresses